CSS/js files are now renamed with their md5 checksums, instead of git checksums
It looks like github will clone a flattened git history when running an action.
So in our case, during the call to build.yml, the individual commits on any given
css file is flattened into the latest commit, repository-wise.

This means that all css files will be renamed each time there is a build, which
is undesirable due to the noise it creates in each diff.

Renaming on the checkums of the actual contents of the file should prevent this.

Javascript files are now also included in this scheme.

To save some template calculations, the generated files are stored in the
environment variable 'filekeys'